% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addBinExtension}
\alias{addBinExtension}
\title{Extends TE bins}
\usage{
addBinExtension(bins, downstream = 1000, binsize = 20,
  min_dist = 100, tss = GRanges(), polyA = GRanges())
}
\arguments{
\item{bins}{Data.frame. TE bin intervals and annotation.}

\item{downstream}{Integer. Basepairs to extend downstream.}

\item{binsize}{Integer. Tile size of extended intervals.}
}
\description{
Extends TE bins
}
