% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addCounts}
\alias{addCounts}
\title{Add TE/Gene counts to scSet}
\usage{
addCounts(scSet, bams = NULL, msa_dir = NULL,
  te_exon_overlap = "filter", hierarchy = "locus", bin_size = NA,
  min_cell_size = 100, use_gcluster = FALSE)
}
\description{
Add TE/Gene counts to scSet
}
