% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calcBackground}
\alias{calcBackground}
\title{Genes should be curated!}
\usage{
calcBackground(BSgenome = Hsapiens, mappable = NULL, tes, genes, reads,
  min_cov = 1, bin_size = 100)
}
\description{
Genes should be curated!
}
