% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compCounts}
\alias{compCounts}
\title{Counts number of reads per interval and cell barcode}
\usage{
compCounts(reads, tes, genes, bin_size = 20, protocol = "fiveprime")
}
\description{
Counts number of reads per interval and cell barcode
}
