% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{repnames}
\alias{repnames}
\title{Returns transposable element families from scSet}
\usage{
repnames(scSet, pattern = NULL, fixed = FALSE)
}
\arguments{
\item{scSet}{scSet object to query.}

\item{pattern}{Regular expression. Returns only families that match \code{pattern}.}

\item{fixed}{Logical. If \code{TRUE}, \code{pattern} is matched as is.}
}
\value{
Character vector.
}
\description{
Retrieves all available families in the provided scSet that match defined characteristics.
}
\examples{
\dontrun{
gs <- createscSet(genome = Hsapiens, tes = te_annotation_df)
families <- names(gs, pattern = 'LTR12') # returns all families containing 'LTR12' in their name.

# Using a pattern not found in scSet will suggest similar hits, e.g.:
names(gs, pattern = 'LTR5HS')
}
}
\seealso{
\code{\link[=createscSet]{createscSet()}}
}
