% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSet.R
\docType{class}
\name{scSet-class}
\alias{scSet-class}
\alias{scSet}
\title{An S4 class to store genomic annotations and results of Repguide functions.}
\description{
An S4 class to store genomic annotations and results of Repguide functions.
}
\section{Slots}{

\describe{
\item{\code{genome}}{BSgenome object.}

\item{\code{tes}}{GRanges object}

\item{\code{cis}}{GRanges object}

\item{\code{refdir}}{Character}

\item{\code{tempdir}}{Character}

\item{\code{targets}}{GRanges object}

\item{\code{blacklist}}{GRanges object}

\item{\code{whitelist}}{GRanges object}

\item{\code{plots}}{List}

\item{\code{guide_length}}{Integer}

\item{\code{families}}{Character}

\item{\code{PAM}}{Character}

\item{\code{kmers}}{GRanges}

\item{\code{alignments}}{DNAStringSetList}

\item{\code{consensus}}{DNAStringSet}

\item{\code{combinations}}{Data.frame}

\item{\code{calls}}{List}

\item{\code{n_cores}}{Integer}

\item{\code{seed}}{Integer}
}}

