% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_select.R
\name{selectCells}
\alias{selectCells}
\title{Call cell barcodes}
\usage{
selectCells(scSet, method = "size", min_size = 0, max_mito = 1,
  min_ribo = 0, lower = 100, niters = 10000, fdr = 0.01)
}
\arguments{
\item{method}{Character. 'Emptydrops' (the default) or 'size' are currently supported.}

\item{min_size}{Integer. Size parameter, minimal number of molecules per cell.}

\item{lower}{Integer. Emptydrops parameter, see}

\item{niters}{Integer. Emptydrops parameter, see}

\item{fdr}{Numeric. Emptydrops parameter, see}
}
\description{
Call cell barcodes
}
