% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_select.R
\name{selectFeatures}
\alias{selectFeatures}
\title{Adds gene/te stats to scSet}
\usage{
selectFeatures(scSet, ds = NULL, class = c("DNA", "SINE", "LINE",
  "LTR"), peak_only = FALSE, reg = 10, min_expr = 25,
  min_expr_third = 3, min_norm_var = 0.3, k = NULL)
}
\description{
Adds gene/te stats to scSet
}
