% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_select.R
\name{selectPeaks}
\alias{selectPeaks}
\title{Selects enriched signal on consensus models}
\usage{
selectPeaks(scSet, neighb_bins = 10, offset = 4, min_expr = 10,
  min_fc = 1, fdr = 0.01, min_loci = 3, summarize = FALSE)
}
\description{
Selects enriched signal on consensus models
}
