% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compCounts}
\alias{compCounts}
\title{Counts number of reads per interval and cell barcode}
\usage{
compCounts(bams, tes, tes_3p, genes, hierarchy = "locus",
  bin_size = NA, sense_only = TRUE, conversion = NULL,
  protocol = "fiveprime", use_gcluster = FALSE)
}
\description{
Counts number of reads per interval and cell barcode
}
